pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sms"
version = "0.0.1"
authors = [
  { name="Matheus Horstmann", email="me@math.eus" },
]
description = "A SMS relay, it is triggered by gammu"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "python-telegram-bot>=21.9",
]

[project.optional-dependencies]
dev = [
  "black>=23.11.0",
  "flake8>=6.1.0",
  "isort>=5.12.0",
  "pre-commit>=3.5.0",
  "pylint>=3.0.2",
  "pyenchant==3.2.2",
  "semantic-version==2.10.0",
  "nltk>=3.9.1",
]
plot = [
  "matplotlib>=3.9.2",
  "PyQt6>=6.7.1"
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 79
target-version = ['py37', 'py38']
include = '\.pyi?$'


[tool.isort]
atomic = true
profile = "black"
line_length = 79
skip_gitignore = true

[tool.pylint.main]

jobs = 2
fail-under = 9

# Files or directories to be skipped. They should be base names, not paths.
ignore = [
  ".venv",
  "README.md",
  "pyproject.toml",
  "VERSION",
]
ignore-patterns="(.)*.(txt|config|cfg|sh|toml|yaml|yml|MD),.gitignore"
persistent = true
py-version = "3.8"
suggestion-mode = true


[tool.pylint.format]
# Maximum number of characters on a single line.
max-line-length = 79

# [tool.pylint.spelling]
# en_CA (aspell), en_GB (aspell), en_US (aspell).
# spelling-dict = "en_GB"
# spelling-ignore-words = []

[tool.pylint."messages control"]
disable = [
  "raw-checker-failed",
  "bad-inline-option",
  "locally-disabled",
  "file-ignored",
  "suppressed-message",
  "useless-suppression",
  "deprecated-pragma",
  "use-implicit-booleaness-not-comparison-to-string",
  "use-implicit-booleaness-not-comparison-to-zero",
  "use-symbolic-message-instead",
  "fixme",
  "missing-function-docstring",
  "missing-module-docstring",
  "missing-class-docstring",
  "pointless-string-statement",
  "no-name-in-module", # Temp
]
